+++ packages/address/src/graphqlQuery.ts
@@ -3,15 +3,25 @@ query countries($locale: SupportedLocale!) {
   countries(locale: $locale) {
     name
     code
+    continent
     phoneNumberPrefix
-    address2Key
-    provinceKey
-    zipKey
+    labels {
+      address1
+      address2
+      city
+      company
+      country
+      firstName
+      lastName
+      phone
+      postalCode
+      zone
+    }
     formatting {
       edit
       show
     }
-    provinces {
+    zones {
       name
       code
     }
@@ -22,15 +32,25 @@ query country($countryCode: SupportedCountry!, $locale: SupportedLocale!) {
   country(countryCode: $countryCode, locale: $locale) {
     name
     code
+    continent
     phoneNumberPrefix
-    address2Key
-    provinceKey
-    zipKey
+    labels {
+      address1
+      address2
+      city
+      company
+      country
+      firstName
+      lastName
+      phone
+      postalCode
+      zone
+    }
     formatting {
       edit
       show
     }
-    provinces {
+    zones {
       name
       code
     }