+++ packages/react-i18n/src/i18n.ts
@@ -15,11 +15,7 @@ import {
   RTL_LANGUAGES,
   Weekdays,
 } from './constants';
-import {
-  MissingCurrencyCodeError,
-  MissingTimezoneError,
-  MissingCountryError,
-} from './errors';
+import {MissingCurrencyCodeError, MissingCountryError} from './errors';
 import {
   getCurrencySymbol,
   translate,
@@ -164,10 +160,15 @@ export default class I18n {
   ): string {
     const {locale, defaultTimezone: timezone} = this;
 
-    if (timezone == null && (options == null || options.timeZone == null)) {
-      throw new MissingTimezoneError(
-        `No timezone code provided. formatDate() cannot be called without a timezone.`,
-      );
+    // Etc/GMT+12 is not supported in most browsers and there is no equivalent fallback
+    if (
+      options &&
+      options.timeZone != null &&
+      options.timeZone === 'Etc/GMT+12'
+    ) {
+      const adjustedDate = new Date(date.valueOf() - 12 * 60 * 60 * 1000);
+
+      return this.formatDate(adjustedDate, {...options, timeZone: 'UTC'});
     }
 
     const {style = undefined, ...formatOptions} = options || {};